Distinguish absolute/relative Seeker, rename, document
